requirements.txt
geopy==2.3.0
googletrans==3.0.0
numpy==1.19.1
Pillow==8.4.0
Pillow==9.5.0
pyttsx3==2.90
Requests==2.30.0
streamlit==1.22.0
tensorflow==2.12.0
